restructured audio and video processor into separate services

+++ video_processing/recorder.py
@@ -6,47 +6,28 @@ import os
 import numpy as np
 import cv2
 from moviepy.editor import *
-#from video_cartoonizer.cartoonizer import caart
 
+class VidRecorder:
 
-class WaveRecorder:
-
-    def __init__(self, filename, sample_rate, depth, channels):
-        self.dat_filename = filename + '.dat'
-        self.wav_filename = filename + '.wav'
+    def __init__(self, filename,aud_filename,vid_img_dir, is_save,sample_rate,depth,channels):
+        self.vid_filename = filename + '.webm'
+        self.cart_vid = filename+'_caart'+'.mp4'
+        self.img_file = filename
+        self.closed = False
+        self.save_video = is_save
+        self.vid_img_dir = vid_img_dir
+        self.dat_filename = aud_filename + '.dat'
+        self.wav_filename = aud_filename + '.wav'
         self.sample_rate = sample_rate
         self.depth = depth
         self.channels = channels
-        self.closed = False
 
-    def write(self, data):
+    def write_audio(self, data):
         try:
             with open(self.dat_filename, "ab") as f:
                 f.write(data)
         except Exception as e:
             print('Unable to write data to file: {0}'.format(e))
-
-    def close(self):
-        if not self.closed:
-            with open(self.dat_filename, "rb") as f:
-                all_data = f.read()
-                all_data = np.frombuffer(all_data, np.float32 if self.depth == 4 else np.int16, -1)
-                chunk_length = int(len(all_data) / self.channels)
-                all_data = np.reshape(all_data, (chunk_length, self.channels))
-            wavfile.write(self.wav_filename, self.sample_rate, all_data)
-            os.remove(self.dat_filename)
-            self.closed = True
-
-class VidRecorder:
-
-    def __init__(self, filename,sample_rate,depth,channels):
-        self.vid_filename = filename + '.webm'
-        self.cart_vid = filename+'_caart'+'.mp4'
-        self.img_file = filename
-        self.closed = False
-        self.sample_rate = sample_rate
-        self.depth = depth
-        self.channels = channels
 
     def write(self, data):
         try:
@@ -63,10 +44,19 @@ class VidRecorder:
         except Exception as e:
             print('Unable to read temp audio data to file: {0}'.format(e))  
 
+    def convert_dat_to_wav(self):
+        with open(self.dat_filename, "rb") as f:
+            all_data = f.read()
+            all_data = np.frombuffer(all_data, np.float32 if self.depth == 4 else np.int16, -1)
+            chunk_length = int(len(all_data) / self.channels)
+            all_data = np.reshape(all_data, (chunk_length, self.channels))
+        wavfile.write(self.wav_filename, self.sample_rate, all_data)
+        os.remove(self.dat_filename)
+
     def close(self):
         if not self.closed:
             try:
-                if os.path.isfile(self.vid_filename):
+                if not self.save_video and  os.path.isfile(self.vid_filename):
                     os.remove(self.vid_filename)        
             except Exception as e:
               logging.info('Unable to delete video file: {0}'.format(e))
